+++ src/commands/utils.js
@@ -28,7 +28,7 @@ export const createButton = (id, label, style, emoji) => {
  */
 export const buttonHandler = (interaction) => {
   const { customId } = interaction;
-  if (customId === "gift") giftButtonHandler(interaction);
+  if (customId.startsWith("gift")) giftButtonHandler(interaction);
   else if (customId.startsWith("announce")) announceButtonHandler(interaction);
   else if (customId.startsWith("eventRole"))
     eventRolesButtonHandler(interaction);